v2 ready

+++ essentials-uninstaller/src/main/java/org/onehippo/cms7/essentials/plugins/uninstaller/UnInstallResource.java
@@ -1,10 +1,7 @@
 package org.onehippo.cms7.essentials.plugins.uninstaller;
 
 import com.fasterxml.jackson.jaxrs.json.JacksonJsonProvider;
-import org.apache.commons.io.file.PathUtils;
 import org.apache.commons.vfs2.FileObject;
-import org.apache.commons.vfs2.FileSystemManager;
-import org.apache.commons.vfs2.VFS;
 import org.apache.cxf.helpers.IOUtils;
 import org.apache.cxf.jaxrs.client.JAXRSClientFactory;
 import org.apache.cxf.jaxrs.client.WebClient;
@@ -14,6 +11,7 @@ import org.onehippo.cms7.essentials.plugin.sdk.instruction.parser.DefaultInstruc
 import org.onehippo.cms7.essentials.plugin.sdk.utils.TemplateUtils;
 import org.onehippo.cms7.essentials.plugins.uninstaller.model.SimpleNode;
 import org.onehippo.cms7.essentials.plugins.uninstaller.model.UninstallModel;
+import org.onehippo.cms7.essentials.plugins.uninstaller.service.MavenExtendedDependencyService;
 import org.onehippo.cms7.essentials.plugins.uninstaller.service.PluginService;
 import org.onehippo.cms7.essentials.plugins.uninstaller.service.PluginSet;
 import org.onehippo.cms7.essentials.sdk.api.install.Instruction;
@@ -23,22 +21,17 @@ import org.onehippo.cms7.essentials.sdk.api.model.rest.UserFeedback;
 import org.onehippo.cms7.essentials.sdk.api.service.*;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
+import org.springframework.context.ApplicationContext;
 
 import javax.inject.Inject;
-import javax.jcr.*;
-import javax.jcr.lock.LockException;
-import javax.jcr.nodetype.ConstraintViolationException;
-import javax.jcr.nodetype.NoSuchNodeTypeException;
-import javax.jcr.version.VersionException;
-import javax.servlet.http.HttpServletResponse;
+import javax.jcr.RepositoryException;
+import javax.jcr.Session;
 import javax.ws.rs.*;
-import javax.ws.rs.core.Context;
 import javax.ws.rs.core.MediaType;
 import javax.ws.rs.core.Response;
 import javax.xml.bind.JAXBContext;
 import javax.xml.bind.JAXBException;
 import javax.xml.bind.Unmarshaller;
-import java.io.File;
 import java.io.IOException;
 import java.io.InputStream;
 import java.net.URL;
@@ -55,6 +48,12 @@ public class UnInstallResource {
 
     private static final Logger log = LoggerFactory.getLogger(UnInstallResource.class);
 
+//
+//    @Inject
+//    MavenExtendedDependencyService mavenExtendedDependencyService;
+
+    @Inject
+    private ApplicationContext appContext;
     @Inject
     private JcrService jcrService;
     @Inject
@@ -143,14 +142,14 @@ public class UnInstallResource {
                 .forEach(allPluginSet::add);
 
         Set<PluginDescriptor> inheritedPlugins = new HashSet<>();
-        allPlugins.stream().forEach(pluginDescriptor -> Optional.ofNullable(
-                pluginDescriptor.getPluginDependencies())
-                .ifPresent(
-                        dependencies -> inheritedPlugins
-                                .addAll(dependencies.stream()
-                                        .map(dependency -> allPluginSet.getPlugin(dependency.getPluginId()))
-                                        .filter(Objects::nonNull)
-                                        .collect(Collectors.toList()))));
+//        allPlugins.stream().forEach(pluginDescriptor -> Optional.ofNullable(
+//                pluginDescriptor.getPluginDependencies())
+//                .ifPresent(
+//                        dependencies -> inheritedPlugins
+//                                .addAll(dependencies.stream()
+//                                        .map(dependency -> allPluginSet.getPlugin(dependency.getPluginId()))
+//                                        .filter(Objects::nonNull)
+//                                        .collect(Collectors.toList()))));
 
         List<PluginDescriptor> installedPlugins = allPlugins.stream()
                 .filter(pluginDescriptor -> Arrays.asList(InstallState.INSTALLED, InstallState.INSTALLING, InstallState.INSTALLATION_PENDING).contains(pluginDescriptor.getState()))
@@ -170,9 +169,11 @@ public class UnInstallResource {
     @SuppressWarnings("DuplicateBranchesInSwitch")
     @GET
     @Path("/plugin/{plugin}/uninstall")
-    public void uninstall(@PathParam("plugin") String plugin) throws IOException {
+    public List<UserFeedback> uninstall(@PathParam("plugin") String plugin) throws IOException {
         UninstallModel uninstallModel = getUninstallModel(plugin);
         List<Instruction> instructions = uninstallModel.getInstructions();
+
+        List<UserFeedback> feedbackMessages = new ArrayList<>();
 
         instructions.forEach(instruction -> {
 
@@ -184,22 +185,38 @@ public class UnInstallResource {
                         String targetPom = mavenDependencyInstruction.getTargetPom();
                         String pluginLocation = "/META-INF/maven." + mavenDependencyInstruction.getGroupId() + "." + mavenDependencyInstruction.getArtifactId() + "/pom.xml";
                         URL resource = getClass().getResource(pluginLocation);
+//                        ApplicationContext.
+//                        InputStream resourceAsStream = ServletContext.getResourceAsStream("/srPaySlipSalaryComp.jasper");
+//                        ApplicationContextProvider.get
+
+//                        final Module module = Module.pomForName(targetPom);
+
+//                        mavenExtendedDependencyService.re
+                        feedbackMessages.add(new UserFeedback().addError("manual step required, remove "
+                                + mavenDependencyInstruction.getGroupId() + ":" + mavenDependencyInstruction.getArtifactId() +
+                                " from " + mavenDependencyInstruction.getTargetPom()));
                         break;
                     case XmlInstruction:
                         XmlInstruction xmlInstruction = InstructionType.getInstruction(instruction);
                         final String xmlTarget = TemplateUtils.replaceTemplateData(xmlInstruction.getTarget(), uninstallModel.getParameters());
                         String nodeName = getName(xmlInstruction.getSource());
-                        removeNode(xmlTarget, nodeName);
+                        if (XmlInstruction.Action.COPY.equals(xmlInstruction.getActionEnum())) {
+                            removeNode(xmlTarget, nodeName);
+                        }
                         break;
                     case FileInstruction:
                         FileInstruction fileInstruction = InstructionType.getInstruction(instruction);
                         final String fileTarget = TemplateUtils.replaceTemplateData(fileInstruction.getTarget(), uninstallModel.getParameters());
-                        removeFile(fileTarget);
+                        if (FileInstruction.Action.COPY.equals(fileInstruction.getActionEnum())) {
+                            removeFile(fileTarget);
+                        }
                         break;
                     case FreemarkerInstruction:
                         FileInstruction freeMarkerInstruction = InstructionType.getInstruction(instruction);
                         final String freeMarkerTarget = TemplateUtils.replaceTemplateData(freeMarkerInstruction.getTarget(), uninstallModel.getParameters());
-                        removeFile(freeMarkerTarget);
+                        if (FreemarkerInstruction.Action.COPY.equals(freeMarkerInstruction.getActionEnum())) {
+                            removeFile(freeMarkerTarget);
+                        }
                         break;
                     default:
                         break;
@@ -207,13 +224,18 @@ public class UnInstallResource {
             } catch (IllegalArgumentException e) {
 
             }
-
-
         });
 
-        String essentialsRoot = (String) uninstallModel.getParameters().get("essentialsRoot");
-        String pluginPath = Paths.get(essentialsRoot, "src/main/resources", plugin + ".xml").toString();
-        removeFile(pluginPath);
+        if (feedbackMessages.isEmpty()) {
+            String essentialsRoot = (String) uninstallModel.getParameters().get("essentialsRoot");
+            String pluginPath = Paths.get(essentialsRoot, "src/main/resources", plugin + ".xml").toString();
+            removeFile(pluginPath);
+            feedbackMessages.add(new UserFeedback().addSuccess(plugin + " uninstalled"));
+        } else {
+            feedbackMessages.add(new UserFeedback().addError(plugin + " partly uninstalled, check remaining steps"));
+        }
+
+        return feedbackMessages;
     }
 
 
@@ -225,7 +247,7 @@ public class UnInstallResource {
         try {
             Files.deleteIfExists(Paths.get(path));
         } catch (IOException e) {
-            e.printStackTrace();
+            log.warn("file does not exist anymore, all good proceeding");
         }
     }
 
@@ -235,7 +257,7 @@ public class UnInstallResource {
             session.getNode(path).getNode(name).remove();
             session.save();
         } catch (RepositoryException e) {
-            e.printStackTrace();
+            log.warn("node does not exist anymore, all good proceeding");
         } finally {
             session.logout();
         }
@@ -248,7 +270,7 @@ public class UnInstallResource {
             SimpleNode simpleNode = (SimpleNode) unmarshaller.unmarshal(getClass().getResourceAsStream("/" + source));
             return simpleNode.getName();
         } catch (JAXBException e) {
-            e.printStackTrace();
+            log.warn("error parsing through the jcr xml, cannot get name");
         }
         return null;
     }
@@ -267,7 +289,7 @@ public class UnInstallResource {
             this.simpleClassName = simpleClassName;
         }
 
-        public static <T> T getInstruction(Instruction instruction) {
+        public static <T extends Instruction> T getInstruction(Instruction instruction) {
             return (T) instruction;
         }
     }
